kucoin_manager/web/application.py
from importlib import metadata

from fastapi import FastAPI
from fastapi.responses import UJSONResponse
from fastapi.staticfiles import StaticFiles
from tortoise.contrib.fastapi import register_tortoise

from kucoin_manager.db.config import TORTOISE_CONFIG
from kucoin_manager.web.api.router import api_router
from kucoin_manager.web.lifetime import register_shutdown_event, register_startup_event


def get_app() -> FastAPI:
    """
    Get FastAPI application.

    This is the main constructor of an application.

    :return: application.
    """
    app = FastAPI(
        title="kucoin_manager",
        description="Manage your kucoin accounts and transactions",
        version=metadata.version("kucoin_manager"),
        docs_url="/api/docs",
        redoc_url="/api/redoc",
        openapi_url="/api/openapi.json",
        default_response_class=UJSONResponse,
    )

    # Adds startup and shutdown events.
    register_startup_event(app)
    register_shutdown_event(app)

    # Main router for the API.
    app.include_router(router=api_router, prefix="")
    # Configures tortoise orm.
    register_tortoise(
        app,
        config=TORTOISE_CONFIG,
        add_exception_handlers=True,
        generate_schemas=True,
    )

    app.mount("/static", StaticFiles(directory="kucoin_manager/static"), name="static")

    return app
